Replace numpy.bool with numpy.bool_
In older numpy releases, numpy.bool was the same as Python bool. This was deprecated in 1.20 and later removed for 1.x, then the name numpy.bool was reintroduced in numpy 2.x as an alias for numpy.bool_, a Boolean value stored as a byte. By using numpy.bool_, we get the same data type on all Python versions.